Move packing to TRImageControl
