+++ TRDataControl/Transport/TR1/TR1DataImporter.cs
@@ -1,10 +1,10 @@
 ﻿using Newtonsoft.Json;
+using TRImageControl.Packing;
 using TRLevelControl.Model;
 using TRModelTransporter.Data;
 using TRModelTransporter.Handlers;
 using TRModelTransporter.Handlers.Textures;
 using TRModelTransporter.Model.Definitions;
-using TRModelTransporter.Model.Textures;
 using TRTexture16Importer.Helpers;
 
 namespace TRModelTransporter.Transport;
