+++ TRDataControl/Transport/TR2/TR2DataExporter.cs
@@ -1,8 +1,8 @@
-﻿using TRLevelControl.Model;
+﻿using TRImageControl.Packing;
+using TRLevelControl.Model;
 using TRModelTransporter.Data;
 using TRModelTransporter.Handlers;
 using TRModelTransporter.Model.Definitions;
-using TRModelTransporter.Model.Textures;
 
 namespace TRModelTransporter.Transport;
