+++ TRDataControl/Transport/TR2/TR2DataImporter.cs
@@ -1,9 +1,9 @@
 ﻿using Newtonsoft.Json;
+using TRImageControl.Packing;
 using TRLevelControl.Model;
 using TRModelTransporter.Data;
 using TRModelTransporter.Handlers;
 using TRModelTransporter.Model.Definitions;
-using TRModelTransporter.Model.Textures;
 
 namespace TRModelTransporter.Transport;
