+++ TRDataControl/Transport/TRDataExporter.cs
@@ -1,10 +1,10 @@
 ﻿using Newtonsoft.Json;
 using System.Drawing;
 using System.Drawing.Imaging;
+using TRImageControl.Packing;
 using TRModelTransporter.Events;
 using TRModelTransporter.Handlers;
 using TRModelTransporter.Model;
-using TRModelTransporter.Model.Textures;
 using TRModelTransporter.Utilities;
 using TRTexture16Importer.Textures;
