+++ TRDataControl/Transport/TRDataImporter.cs
@@ -1,6 +1,6 @@
-﻿using TRModelTransporter.Handlers;
+﻿using TRImageControl.Packing;
+using TRModelTransporter.Handlers;
 using TRModelTransporter.Model;
-using TRModelTransporter.Model.Textures;
 
 namespace TRModelTransporter.Transport;
