+++ TRDataControl/Utilities/AbstractMassTRTextureDeduplicator.cs
@@ -2,7 +2,6 @@
 using System.Diagnostics;
 using System.Drawing.Imaging;
 using TRImageControl.Packing;
-using TRModelTransporter.Model.Textures;
 
 namespace TRModelTransporter.Utilities;
