+++ TRDataControl/Utilities/TRTextureClassifier.cs
@@ -1,6 +1,6 @@
 ﻿using System.Security.Cryptography;
 using System.Text;
-using TRModelTransporter.Model.Textures;
+using TRImageControl.Packing;
 
 namespace TRModelTransporter.Utilities;
