+++ TRImageControl/Packing/PackingException.cs
