+++ TRImageControl/Packing/Textures/PositionedTexture.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace TRModelTransporter.Model.Textures;
+namespace TRImageControl.Packing;
 
 public class PositionedTexture
 {
