+++ TRImageControl/Packing/Textures/RemapTypes/TR1TextureRemapGroup.cs
@@ -1,7 +1,6 @@
-﻿using TRImageControl.Packing;
-using TRLevelControl.Model;
+﻿using TRLevelControl.Model;
 
-namespace TRModelTransporter.Model.Textures;
+namespace TRImageControl.Packing;
 
 public class TR1TextureRemapGroup : AbstractTextureRemapGroup<TR1Type, TR1Level>
 {
