+++ TRImageControl/Packing/Textures/RemapTypes/TR2TextureRemapGroup.cs
@@ -1,7 +1,6 @@
-﻿using TRImageControl.Packing;
-using TRLevelControl.Model;
+﻿using TRLevelControl.Model;
 
-namespace TRModelTransporter.Model.Textures;
+namespace TRImageControl.Packing;
 
 public class TR2TextureRemapGroup : AbstractTextureRemapGroup<TR2Type, TR2Level>
 {
