+++ TRImageControl/Packing/Textures/TextureDependency.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace TRModelTransporter.Model.Textures;
+namespace TRImageControl.Packing;
 
 public class TextureDependency<E> where E : Enum
 {
