+++ TRImageControl/Packing/Types/TR2TexturePacker.cs
@@ -1,22 +1,17 @@
 ﻿using System.Drawing;
 using TRLevelControl;
 using TRLevelControl.Model;
-using TRModelTransporter.Helpers;
-using TRModelTransporter.Model.Textures;
 using TRTexture16Importer;
-using TRTexture16Importer.Helpers;
 
 namespace TRImageControl.Packing;
 
-public class TR1TexturePacker : AbstractTexturePacker<TR1Type, TR1Level>
+public class TR2TexturePacker : AbstractTexturePacker<TR2Type, TR2Level>
 {
     private const int _maximumTiles = 16;
 
-    public TRPalette8Control PaletteManager { get; set; }
-
     public override int NumLevelImages => Level.Images8.Count;
 
-    public TR1TexturePacker(TR1Level level, ITextureClassifier classifier = null)
+    public TR2TexturePacker(TR2Level level, ITextureClassifier classifier = null)
         : base(level, _maximumTiles, classifier) { }
 
     protected override List<AbstractIndexedTRTexture> LoadObjectTextures()
@@ -25,7 +20,7 @@ public class TR1TexturePacker : AbstractTexturePacker<TR1Type, TR1Level>
         for (int i = 0; i < Level.ObjectTextures.Count; i++)
         {
             TRObjectTexture texture = Level.ObjectTextures[i];
-            if (texture.IsValid())
+            //if (texture.IsValid())
             {
                 textures.Add(new IndexedTRObjectTexture
                 {
@@ -45,7 +40,7 @@ public class TR1TexturePacker : AbstractTexturePacker<TR1Type, TR1Level>
         for (int i = 0; i < sprites.Count; i++)
         {
             TRSpriteTexture texture = sprites[i];
-            if (texture.IsValid())
+            //if (texture.IsValid())
             {
                 textures.Add(new IndexedTRSpriteTexture
                 {
@@ -58,25 +53,27 @@ public class TR1TexturePacker : AbstractTexturePacker<TR1Type, TR1Level>
         return textures;
     }
 
-    protected override List<TRMesh> GetModelMeshes(TR1Type modelEntity)
+    protected override List<TRMesh> GetModelMeshes(TR2Type modelEntity)
     {
         return Level.Models[modelEntity]?.Meshes;
     }
 
-    protected override TRSpriteSequence GetSpriteSequence(TR1Type entity)
+    protected override TRSpriteSequence GetSpriteSequence(TR2Type entity)
     {
         return Level.Sprites[entity];
     }
 
-    protected override IEnumerable<TR1Type> GetAllModelTypes()
+    protected override IEnumerable<TR2Type> GetAllModelTypes()
     {
         return Level.Models.Keys.ToList();
     }
 
     protected override void CreateImageSpace(int count)
     {
+        // We ignore 8-bit images, but the numbers must match
         for (int i = 0; i < count; i++)
         {
+            Level.Images16.Add(new());
             Level.Images8.Add(new()
             {
                 Pixels = new byte[TRConsts.TPageSize]
@@ -86,20 +83,11 @@ public class TR1TexturePacker : AbstractTexturePacker<TR1Type, TR1Level>
 
     public override Bitmap GetTile(int tileIndex)
     {
-        return Level.Images8[tileIndex].ToBitmap(Level.Palette);
+        return Level.Images16[tileIndex].ToBitmap();
     }
 
     public override void SetTile(int tileIndex, Bitmap bitmap)
     {
-        PaletteManager ??= new()
-        {
-            Level = Level
-        };
-        PaletteManager.ChangedTiles[tileIndex] = bitmap;
-    }
-
-    protected override void PostCommit()
-    {
-        PaletteManager?.MergeTiles();
+        Level.Images16[tileIndex].Pixels = TextureUtilities.ImportFromBitmap(bitmap);
     }
 }
