+++ TRImageControl/TRImageControl.csproj
@@ -13,4 +13,9 @@
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="System.Drawing.Common" Version="8.0.4" />
   </ItemGroup>
+  <ItemGroup>
+    <Reference Include="RectanglePacker">
+      <HintPath>..\Deps\RectanglePacker.dll</HintPath>
+    </Reference>
+  </ItemGroup>
 </Project>
