+++ TRRandomizerCore/Textures/Wireframing/TR3Wireframer.cs
@@ -4,7 +4,6 @@ using TRImageControl.Packing;
 using TRLevelControl.Helpers;
 using TRLevelControl.Model;
 using TRModelTransporter.Helpers;
-using TRModelTransporter.Model.Textures;
 using TRRandomizerCore.Utilities;
 using TRTexture16Importer.Helpers;
