+++ TextureExport/Types/DependencyExporter.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
+using TRImageControl.Packing;
 using TRLevelControl.Model;
-using TRModelTransporter.Model.Textures;
 
 namespace TextureExport.Types;
